Modift: 1. modify data type from 'int' to 'Long' in ListComputer('sum' and 'sum_x2');    int sum => long sum;    int sum_x2 => long sum_x2; 2. select insert to save, delete restrictions of qual;    delete (record.qual < minqual);

+++ src/main/java/org/bgi/flexlab/gaea/tools/callsv/ListComputer.java
@@ -17,7 +17,7 @@ public class ListComputer {
 		if(l.size() == 0) {
 			return 0;
 		}
-		int sum = 0;
+		long sum = 0;
 		for(Iterator<Integer> insertlist = l.iterator(); insertlist.hasNext();) {
 			sum += insertlist.next();
 		}
@@ -47,7 +47,7 @@ public class ListComputer {
 			return 0;
 		}
 		
-		int sum_x2 = 0;
+		long sum_x2 = 0;
 		for(Iterator<Integer> insertlist = l.iterator(); insertlist.hasNext();) {
 			sum_x2 += Math.pow((insertlist.next() - mean), 2);
 		}
